[Refactor] #37: config 및 jwt 파일 이동
