+++ src/main/java/com/example/off/common/config/FilterConfig.java
@@ -1,7 +1,7 @@
-package com.example.off.config;
+package com.example.off.common.config;
 
-import com.example.off.jwt.JwtAuthenticationFilter;
-import com.example.off.jwt.JwtTokenProvider;
+import com.example.off.common.jwt.JwtAuthenticationFilter;
+import com.example.off.common.jwt.JwtTokenProvider;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
