+++ src/main/java/com/example/off/common/config/PasswordConfig.java
@@ -1,4 +1,4 @@
-package com.example.off.config;
+package com.example.off.common.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
