+++ src/main/java/com/example/off/common/interceptor/StompAuthChannelInterceptor.java
@@ -1,8 +1,7 @@
 package com.example.off.common.interceptor;
 
 import com.example.off.common.auth.StompPrincipal;
-import com.example.off.common.exception.OffException;
-import com.example.off.common.jwt.service.JwtTokenService;
+import com.example.off.common.jwt.JwtTokenService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.Message;
@@ -12,9 +11,6 @@ import org.springframework.messaging.simp.stomp.StompHeaderAccessor;
 import org.springframework.messaging.support.ChannelInterceptor;
 import org.springframework.messaging.support.MessageHeaderAccessor;
 import org.springframework.stereotype.Component;
-
-import static com.example.off.common.response.ResponseCode.INVALID_TOKEN;
-import static com.example.off.common.response.ResponseCode.TOKEN_NOT_FOUND;
 
 @Slf4j
 @RequiredArgsConstructor
