--- src/main/java/com/example/off/domain/member/dto/PortfolioRequest.java
@@ -1,12 +1,0 @@
-package com.example.off.domain.member.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-//Todo: 삭제
-//Todo: PortfolioReqeust/Resposne 통합 (ProfileResponse inner class)
-public record PortfolioRequest(
-        String description,
-        String link
-) {}
